requirements.txt
py-trello==0.4.3
ipdb==0.8.1
urwid==1.3.1
SQLAlchemy==1.0.11
trollius==2.0
